@@ -3,7 +3,10 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "node src/application/index.ts"
+    "start": "node dist/application/index.js",
+    "tsc": "tsc",
+    "postinstall": "npm run tsc",
+    "dev": "ts-node-dev — respawn — pretty — transpile-only index.ts"
   },
   "keywords": [],
   "author": "Antonyuk Oleksandr",
@@ -14,7 +17,7 @@
     "connect-mongodb-session": "^3.1.1",
     "csurf": "^1.11.0",
     "express": "^4.17.1",
-    "express-handlebars": "^6.0.1",
+    "express-handlebars": "^6.0.6",
     "express-session": "^1.17.2",
     "hbs": "^4.2.0",
     "mongoose": "^6.0.13",
@@ -22,7 +25,13 @@
     "redis": "^4.0.2"
   },
   "devDependencies": {
-    "typescript": "^4.8.4"
+    "@types/csurf": "^1.11.2",
+    "@types/express": "^4.17.14",
+    "@types/node": "^18.11.13",
+    "@typescript-eslint/parser": "^5.46.0",
+    "ts-node": "^10.9.1",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^4.9.4"
   },
   "repository": {
     "type": "git",
