@@ -48,5 +48,5 @@ router.post('/login', async (req, res) => {
 
 })
 
-
-module.exports = router
+//export const authRoutes = router
+ module.exports = router
